docs(enigma-mocker): added README.md and updated api docs
