+++ apis/enigma-mocker/src/index.js
@@ -1,3 +1,5 @@
 import fromFixture from './from-fixture';
 
-export default { fromFixture };
+export default {
+  fromFixture,
+};
