+++ apis/enigma-mocker/src/mocks/create-session-object-mock.js
@@ -1,0 +1,12 @@
+function CreateSessionObjectMock() {
+  return (props) =>
+    Promise.resolve({
+      on: () => {},
+      once: () => {},
+      getLayout: () => Promise.resolve({}),
+      id: props?.qInfo?.qId ? props.qInfo.qId : `sel - ${+Date.now()}`,
+      ...props,
+    });
+}
+
+export default CreateSessionObjectMock;
