+++ apis/enigma-mocker/src/mocks/get-app-layout-mock.js
@@ -1,0 +1,5 @@
+function GetAppLayoutMock() {
+  return () => Promise.resolve({ id: 'app-layout' });
+}
+
+export default GetAppLayoutMock;
