+++ apis/enigma-mocker/src/mocks/session-mock.js
@@ -1,0 +1,11 @@
+function SessionMock() {
+  return {
+    getObjectApi() {
+      return Promise.resolve({
+        id: `sessapi - ${+Date.now()}`,
+      });
+    },
+  };
+}
+
+export default SessionMock;
